gemini-file-search/requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
